renamed new css modules to .module.css
