+++ src/components/field-validations/field-validations.react.js
@@ -4,7 +4,7 @@ import cx from 'classnames';
 
 import FieldValidation from '../field-validation';
 
-import styles from './field-validations.css';
+import styles from './field-validations.module.css';
 
 const FieldValidations = props => {
   const {validations, centered, className} = props;
