+++ src/components/fieldset/fieldset.react.js
@@ -7,7 +7,7 @@ import {UIBase} from '../layout';
 import FieldLabel from '../field-label';
 import Text from '../text';
 
-import styles from './fieldset.css';
+import styles from './fieldset.module.css';
 
 export const FieldsetHeader = props => {
   const {legend, description, meta} = props;
