+++ src/components/loading-spinner/loading-spinner.module.css
