+++ src/components/skeleton/components/skeleton-shape.react.js
@@ -4,7 +4,7 @@ import cx from 'classnames';
 
 import {UIBase} from '../../layout';
 
-import styles from '../skeleton.css';
+import styles from '../skeleton.module.css';
 
 const SkeletonShape = props => {
   const {
